ai/requirements..txt
fastapi
uvicorn
openai
python-dotenv
PIL
requests
pydantic
uuid
